requirements.txt
Django==2.0.4
django-markdown-deux==1.0.5
ez-setup==0.9
markdown2==2.3.5
mysql-connector-python==8.0.6
mysqlclient==1.3.12
pytz==2018.4
South==1.0.2
virtualenv==15.2.0
